Refactored helper functions for faceoff generation to be in their own module to keep the model file clean.
